+++ src/CarePackageMod/_FumiKMod.cs
@@ -1,4 +1,5 @@
-﻿using HarmonyLib;
+﻿using Common;
+using HarmonyLib;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -9,9 +10,19 @@ namespace CarePackageMod
 {
     public class FumiKMod : KMod.UserMod2
     {
+        public static Harmony instance;
+
         public override void OnLoad(Harmony harmony)
         {
-            CarePackageState.LoadStrings();
+            instance = harmony;
+            Helpers.ModName = "CustomizeBuildings";
+
+            CustomStrings.LoadStrings();
+#if LOCALE
+            Helpers.StringsPrint();
+#endif
+            Helpers.StringsLoad();
+
             new PeterHan.PLib.Options.POptions().RegisterOptions(this, typeof(CarePackageState));
 
             base.OnLoad(harmony);